requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
requests
python-debian
